+++ Source/Malignant/Mutants/BoneBladeMutantCharacter.cpp
@@ -16,7 +16,7 @@ void ABoneBladeMutantCharacter::BeginPlay()
 void ABoneBladeMutantCharacter::Tick(float DeltaTime)
 {
 	Super::Tick(DeltaTime);
-	//pull request
+
 }
 
 void ABoneBladeMutantCharacter::SetupPlayerInputComponent(UInputComponent* PlayerInputComponent)
