+++ Source/Malignant/Mutants/FireBreathingMutantCharacter.cpp
@@ -1,0 +1,5 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "FireBreathingMutantCharacter.h"
+
