+++ Source/Malignant/Mutants/FireBreathingMutantCharacter.h
@@ -1,0 +1,17 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "../Mutants/MutantCharacter.h"
+#include "FireBreathingMutantCharacter.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class MALIGNANT_API AFireBreathingMutantCharacter : public AMutantCharacter
+{
+	GENERATED_BODY()
+	
+};
