+++ Source/Malignant/Mutants/MutantCharacter.cpp
